requirements.txt
Django==3.2
djangorestframework
pytest
pytest-django
django-filter
factory-boy
django-debug-toolbar